requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-JWT-Extended
python-dotenv
Werkzeug
joblib
numpy
scikit-learn
supabase
